move education to its own component and change order on smaller screens
- TODO clean up shared CSS
