+++ assets/shared-styles.scss
@@ -1,0 +1,73 @@
+@import "~/assets/variables.scss";
+
+.card {
+  margin-top: 2rem;
+  text-align: center;
+}
+
+.container {
+  background-color: none;
+}
+.column {
+  max-width: 400px;
+  min-width: 350px;
+  &-projects {
+    max-width: 600px;
+  }
+}
+.card {
+  margin-top: 2rem;
+  text-align: center;
+}
+.services {
+  background: none;
+  box-shadow: none !important;
+}
+.service {
+  height: 96px;
+  cursor: pointer;
+}
+@media screen and (max-width: 600px) {
+  .service {
+    height: 64px;
+  }
+  .column {
+    min-width: 305px;
+  }
+}
+@media (min-width: 87.5em) {
+}
+.align-text-left {
+  text-align: left;
+}
+
+.short-text {
+  display: flex;
+  justify-content: center;
+}
+.education-row {
+  display: flex;
+  justify-content: left;
+}
+.education-link {
+  text-decoration: none;
+}
+
+.date-spread {
+  display: flex;
+  justify-content: space-between;
+  padding: 0px;
+}
+
+.greyed-out-text {
+  //color: $grey-darken-1;
+  color: $grey-lighten-1;
+}
+
+.link {
+  max-width: 150px;
+  text-overflow: ellipsis;
+  color: white;
+  overflow: hidden;
+  white-space: nowrap;
+}
